requirements.txt
discord.py=1.0.1
sqlalchemy

